+++ ProductScraper.Services/Interfaces/IScrapeConfigService.cs
@@ -7,9 +7,9 @@ namespace ProductScraper.Services.Interfaces
     public interface IScrapeConfigService
     {
         Task<IList<ScrapeConfig>> GetAllAsync();
-        Task<ScrapeConfig> GetDetailsAsync(int id);
+        Task<ScrapeConfig> GetDetailsAsync(long id);
         Task AddAsync(ScrapeConfig scrapeConfig);
         Task UpdateAsync(ScrapeConfig scrapeConfig);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
